+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/model/dto/EventDtos.kt
@@ -559,6 +559,46 @@ internal data class VoteRemovedEventDto(
 ) : ChatEventDto()
 
 @JsonClass(generateAdapter = true)
+internal data class ReminderCreatedEventDto(
+    val type: String,
+    val created_at: ExactDate,
+    val message_id: String,
+    val user_id: String,
+    val cid: String,
+    val reminder: DownstreamReminderDto,
+) : ChatEventDto()
+
+@JsonClass(generateAdapter = true)
+internal data class ReminderUpdatedEventDto(
+    val type: String,
+    val created_at: ExactDate,
+    val message_id: String,
+    val user_id: String,
+    val cid: String,
+    val reminder: DownstreamReminderDto,
+) : ChatEventDto()
+
+@JsonClass(generateAdapter = true)
+internal data class ReminderDeletedEventDto(
+    val type: String,
+    val created_at: ExactDate,
+    val message_id: String,
+    val user_id: String,
+    val cid: String,
+    val reminder: DownstreamReminderDto,
+) : ChatEventDto()
+
+@JsonClass(generateAdapter = true)
+internal data class NotificationReminderDueEventDto(
+    val type: String,
+    val created_at: ExactDate,
+    val message_id: String,
+    val user_id: String,
+    val cid: String,
+    val reminder: DownstreamReminderDto,
+) : ChatEventDto()
+
+@JsonClass(generateAdapter = true)
 internal data class AIIndicatorUpdatedEventDto(
     val type: String,
     val ai_state: String,