+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/repository/ChannelConfigRepositoryTest.kt
@@ -144,6 +144,7 @@ internal class ChannelConfigRepositoryTest {
                 pushNotificationsEnabled = randomBoolean(),
                 uploadsEnabled = randomBoolean(),
                 urlEnrichmentEnabled = randomBoolean(),
+                messageRemindersEnabled = randomBoolean(),
             ),
             emptyList(),
         )