+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/ui/common/feature/channel/info/ChannelInfoViewEvent.kt
@@ -16,13 +16,11 @@
 
 package io.getstream.chat.android.ui.common.feature.channel.info
 
-import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.models.Member
 
 /**
  * Represents side-effect events related to channel information actions.
  */
-@ExperimentalStreamChatApi
 public sealed interface ChannelInfoViewEvent {
 
     /**